move unit test file to test directory
