+++ test/unit/jest-unit.json
@@ -1,0 +1,18 @@
+{
+  "moduleFileExtensions": ["js", "json", "ts"],
+  "rootDir": ".",
+  "testEnvironment": "node",
+  "testRegex": ".spec.ts$",
+  "transform": {
+    "^.+\\.(t|j)s$": "ts-jest"
+  },
+  "collectCoverageFrom": [
+    "**/*.(t|j)s"
+  ],
+  "coverageDirectory": "../coverage",
+  "moduleNameMapper": {
+    "@polkadot/api(.*)$": "<rootDir>../../src/substrate/substrate.handler.ts",
+    "@polkadot/rpc-provider/ws(.*)$": "<rootDir>../../src/substrate/substrate.handler.ts"
+  },
+  "setupFiles": ["dotenv/config"]
+}
